requirements-test.txt
# Testing requirements for LearnSphere

# Core testing framework
pytest==8.4.2               
pytest-cov==7.0.0            
pytest-mock==3.14.1          
pytest-xdist==3.8.0          
pytest-timeout==2.4.0        

# Test utilities
factory-boy>=3.2.0
faker>=18.0.0  
freezegun>=1.2.0  

# Coverage reporting
coverage>=7.0.0  

# Development requirements (included for convenience)
-r requirements.txt  

# Optional: For testing web functionality
pytest-flask>=1.2.0  
selenium>=4.0.0  

# Optional: For performance testing
pytest-benchmark>=4.0.0  

# Optional: For property-based testing
hypothesis>=6.70.0
